Add CIRCT submodule again (our fork)
The submodule is configured to track our fork's "dynamatic" branch.
We'll keep the fork's "main" branch unmodified as to be able to get
upstream changes easily. Whenever we bump CIRCT, we'll rebase the
"dynamatic" branch on "main", potentially resolve conflicts, and then
force push the rebased branch on our fork.