[HandshakeToDot] Prototype conversion pass to DOT
This commit introduces a prototype conversion pass from handshake to dot
that is heavily based on the --handshake-pint-dot pass from CIRCT. Work
remains to make the generated graphs easier to parse by humans (e.g.,
group operations by originating basic block) and/or to make the
generated DOTs interoperable with the old Dynamatic implementation.

As this is our first transformation pass on a dialect, this commits also
introduces the necessary code infrastructure to define and build such
passes.