[FPGA18] Option to identify basic blocks
This commits adds a boolean option to the FPGA18 pass whose effect is to
annotate all operations in each handshake.func of the output IR with an
integer attribute denoting the basic block from which it originates in
the corresponding input func.func. This will be useful, in particular,
to group operations together based on source basic block in the
handshake to DOT conversion pass. This option is off by default (no
attributes generated).

Some unit tests are added to check for correct behavior when the pass is
ran with this option enabled.