[HandshakeToDot] Grouping by basic block
This commit modified the HandshakeToDot pass so that its output format
is closer to the one from the old Dynamatic. Primarily, this means that
operations that used to belong to the same basic block are grouped
together in subgraphs. A utility function/data structure is added to the
StandardToHandshakeFPGA18 pass to aid in that grouping.