Move transformation passes to "Transforms" dir
This commit moves files related to Handshake passes out of the
"Dialect/Handshake" directory and into the "Transforms" directory. In
addition to making the repository's organization more in line with the
LLVM/CIRCT repositories, it also avoids the redefinition of namespaces
for each dialect we touch (e.g., handshake) and the name ambiguities
that ensue in code.