[PushConstants] Constant pushing pass
This commit introduces the --push-constants pass, which moves
arith::ConstantOp operations to the block(s) using them within each
function. This is useful as a preprocessing pass between lowering to
handshake because it reduces the amount of circuitry needed to forward
constants between basic blocks.

The commits also adds simple tests for the pass.