[InferBasicBlocks] Infer BB of handshake ops
This commit introduces a pass that tries to infer the basic block of
untagged operations withing handshake::FuncOps. It does so by
backtracking through the dataflow predecessors of an operation and
using some simple rules to determine whether the operation belongs to
the same block as its predecessors. The core function of this pass is
exposed so that it may be used outside of the pass.

This is meant as a very experimental pass that quickly and pragmatically
solves a simple problem at the moment; inferring the basic block of
forks and sinks after IR materialization so that the HandshakeToDot pass
can display them properly in the block where they belong.

The commit includes a couple basic tests for this pass.