[PrepareForLegacy] Preprocessing path for Handshake
This commit introduces a preprocessing pass for handshake-level IR that
modifies its structure in order to prepare it to be exported to DOT
and used in the legacy version of Dynamatic. At the moment, the only
step the function performs is converting all unconditional branches into
conditional branches with a constant true condition.