[FPGA18] All constants triggered by control merges
This commit changes the policy followed by FPGA18's elastic pass to
connect constants. Instead of them being trigerred by SourceOp
handshake components, they are now triggered by the control merge of
the block in which they are defined. Tests related to the conversion
pass are mmanually modified to reflect that change in behavior.