[HandshakeToNetList] Prototype conversion pass
This commit introduces a prototype handshake-to-netlist pass to
Dynamatic, which converts a handshake function into a "netlist-level"
representation made up of HW instances of external modules
interconnected by ESI channels. As of now the pass attempts to follow
legacy Dynamatic's way of generating RTL designs, in particular the way
in which the circuit connects/interacts with the rest of the system.
This may change in the future once we have defined a formal interface
between the external world and the circuits we generate.