[MemoryAnalysis] Basic memory analysis pass
This commit introduces two passes that form the basis of Dynamatic's
memory analysis infrastructure and that are meant to be ran after
raising the IR to the affine level. At this point, they allow to
identify all dependencies between all pairs of affine memory accesses
in a function, and then store this dependence information in the IR as
attributes attached to memory accesses involved in dependencies.

This is achieved by:
1. Giving a unique name to each memory access (stored as an attribute)
   in the NameMemoryOps pass (--name-memory-ops).
2. Using MLIR's polyhedral analysis infrastructure to identify
   dependencies in the AnalyzeMemoryAccesses pass
   (--analyze-memory-accesses), using those unique names to reference
   specific accesses in stored dependence information.

The goal is to pass this information down to the std-to-handshake
lowering pass that will use it to decide where to place LSQs and simple
MCs in the generated circuits. A future commit will extend the analysis
to work with non-affine memory accesses, which are unsupported at the
moment.