[AffineToScf] New affine lowering pass
This commit introduces a new pass called AffineToScf which is a very
slight variation of the AffineToStandard pass from MLIR. The only
difference is that, when replacing AffineLoadOp's and AffineStoreOp's
with memref::LoadOp's and memref::StoreOp's, respectively, it also
preserves some attributes of interest that the original affine operation
may have had. In practice, this allows us to preserve results of the
memory analysis pass and pass them down the pipeline.