[ExportDOT] Legacy flag & Missing legacy ops
This commits adds a legacy flag to the ExportDOT pass which, when
enabled, instructs the pass to generate a DOT that is compatible with
legacy Dynamatic. This adds some restrictions on the input IR (e.g., no
unconditional branches allowed). By default, the pass will make no
attempt to generate compatible DOTS (e.g., no data annotations will be
added to nodes and edges).

The commit also improves error handling throughout the pass and adds
support for a few previously missing arithmetic operations in legacy
mode.