[ArithReduceArea] Arithmetic optimization pass
This commit introduces a pass that greedily optimizes arithmetic
operations to reduce the area that would be taken up by them on a
circuit. Right now, this is used for the sole purpose of optimizing away
a useless multiplication that is unecessarily genereated in some
situations (e.g., on the fir benchmark).