[Bitwidth] Bitwidth optimization pass #1
This commit introduces a bitwidth optimization pass similar to the one implemented by legacy Dynamatic. The pass iteratively optimizes the bitwidth of all values present in the handshake-level IR by iterating between a forward analysis step (which optimizes the bitwidth of results based on the bitwidth of operands) and a backward analysis step (which optimizes the bitwidth of operands based on the bitwidth of results) until the IR converges to its optimized form.

The commit also introduces two helper passes that are pre-requisites to the bitwidth optimization pass itself. `HandshakeInitIndType` concretizes the width of all `IndexType`'s in the IR while `HandshakeInitCstWidth` modifies the width of all constants to be as narrow as possible.