[OptimizeBitwidth] Fix warnings and delete test
This commit fixes "unused function/variable" warnings in the bitwidth
optimization pass. As a quick and dirty fix (so that we can continue
running `ninja check-dynamatic`) for a probably very real problem, it
also deletes one unit test of the bitwidth optimization pass that would
make the pass run seemingly forever.