Initial support for experimental Dynamatic++ work
This commit adds a new directory at the top level of the repository that
is meant to contain experimental work i.e. work that we want to maintain
on the main branch for visibility and ease of use but that does not
match our standards of code quality and or reliability.