[FrequencyProfiler] Count transition frequencies
This commit intoduces an experimental tool that count transition
frequencies between basic blocks in an std-level function. The tool is
largely inherited from CIRCT's handshake-runner tool with the addition
of a profiler to keep track of the number of transitions on a specific
set of inputs.

The ouput of the tool is useful for the first phase of the smart buffer
placement pass, which requires transition frequencies between basic
blocks. The output may be configured to be CSV-formatted (for
Dynamatic++'s future buffer placement pass) or DOT-formatted (for legacy
Dynamatic's buffer placement pass).