[Tutorial] Source code for first pass tutorial
This commit adds support for creating tutorials in Dynamatic++. It
simply creates a new "tutorials" directory whose internal structure
ressembles closely the structure of the top level directory but which
contains tutorial code (the documentation for which should go in the
"docs" top level directory). The code contained in this directory is
built alongside the rest of the project (a compile-time flag to control
this behavior could be added in the future).

The commit also contains the source code for the first tutorial in
Dynamatic++ (tutorial text to follow) which is about creating a simple
merge-simplification pass using an OpBuilder.