[Tutorial] Source code for greedy pass tutorial
This commit introduces the source code for the second part of the
tutorial about pass creating. It implements a pass that performs almost
exactly the same task as the SimplifyMergeLike pass, but does so using a
greedy pattern rewriter instead of an OpBuilder. It showcases an issue
that arises with the former where a control merge with a single operand
and an unused index result doesn't get deleted from the IR, even though
it could.

The commit also introduces a test file in tutorials/test which will be
useful when writing the tutorial text.