[Support] Logical BB Handling
This commit moves logical basic block handling to a dedicated support
file (instead of being defined as part of the std-to-handshake
conversion pass). This will be useful to centralize the point of
definition of all utilities related to logical BBs, the first of which
is in the inheritBB function which is useful when creating operations
that logically belong to the same BB as their operand(s) producer. Many
passes are modified to adapt to this change in declaration location and
make use of inheritBB.