CIRCT fork: track main branch instead of dynamatic
This commit changes the branch that the project tracks on our CIRCT fork
from dynamatic to main. All commits that were made on the dynamatic
branch of the CIRCT fork were replayed on the main branch, and this
commit makes the superproject track the current HEAD of main.

The dynamatic branch on the CIRCT fork turned out to be unecessary and
added an extra step when bumping CIRCT. The initial idea was to always
keep main clean w.r.t. upstream CIRCT, and rebase the dynamatic branch
on each bump so that we could have a clean history on top of CIRCT's
changes. However, this creates problems as rebase's change commit
hashes, making it impossible to revert the superproject's submodule to
previous commits after a CIRCT bump. The correct way of maintaining the
history is to merge our changes with every CIRCT bump. It is slightly
easier to do that directly on the main branch as it removes the step of
merging to the dynamatic branch. After some thought, I realized we also
don't care about having a "clean" main branch w.r.t. CIRCT. The
dynamatic branch will probably be kept around so that it's possible to
revert to early versions of the project (for the 0 people who will ever
care about such a thing).

So this is as it shall be from now on.