[MinimizeCstWidth] Fix loop value replacement bug
This fixes a bug that would occur whenever narrowing the bitwidth of a
constant operation whose result had multiple users. It was possible for
some users of the constant result to not be updated with the extended
value, causing a type error.

The call to cstOp->getUsers() had to be wrapped into
llvm::make_early_inc_range to ensure that it's safe to iterate on the
list of users while modifyng them. A testcase is added to check for
correct behavior in such a situation.