[CFDFC] Extract CFDFCs from std prof. info (#4)
This commit starts to add support for smart buffer placement by
introducing the CFDFC extraction logic, which uses an MILP and std-level
profiling information to extract CFDFCs from an input dataflow circuit
(handshake::FuncOp). A prototype of the buffer pass that will eventually
place buffers inside our circuits is also included in this commit.

We use the Gurobi optimizer to solve the MILPs that result from our
CFDFC extraction logic. The full code of the pass is built conditionally
based on whether Gurobi is detected on the machine at the time of
running CMake. If it is not detected, a warning is produced during
configuration and the buffer pass simply returns an error.