[Tutorial] Initial work on the MLIR primer
This commit introduces a first and incomplete version of the MLIR
primer, which at this point focuses on high-level IR structure and IR
traversal. It also includes some initial work on describing operations
and their API, though more needs to be done in this area.

The commit also has some typo fixes in the second part of the
"CreatingPasses" tutorial.