[Build] Choose # of threads to build on with ninja
This commit adds an argumetn to the build script (--threads | -t) to
allow the user to choose the number of concurrent threads to build on
with ninja (equivalent to the -j <num_threads> argument in ninja). This
may be helpful on machines that tend to run our of memory when
building/linking LLVM, as it becomes easy to temporarily build on a
low number of threads to give as much memory as possible to critical
steps in the compilation process.