[Tutorial] MLIR primer complete
This commit adds the final sections to the MLIR primer (regions,
blocks, attributes, dialects, printing), which is now in a decent state
for users to start looking at. While it is not meant to be a complete
MLIR reference, small sections may be added to it in the future if it
makes sense (in particular, some more code snippets for common
implementation scenarios may prove to be especially useful to
newcomers).