[MinimizeCstWidth] Do not create duplicated csts
This commit is the first in an effort to improve the handling of
constants in out dataflow circuits. It improves the constant bitwidth
minimization pass by making sure it does not create duplicated constants
(i.e., constants with the same control and value attribute) because of
the minimization process. This is possible if two constants with the
same value initially had different bitwidths higher than the strict
minimum required to fit the value. A couple unit tests are added to
check that the identification and removal of duplicated constants works
as intended.