[FPGA18] Source constants when possible
This commit improves the handling of constants in the FPGA18 elastic
pass. Instead of always being triggered by the control-only network,
constants which are safe to be triggered at all times are now connected
to sources instead, slightly simplifying and improving the performance
of circuits. Constant connection should now follow the same logic as the
elastic pass in legacy Dynamatic.