[ExportDOT] Tool for exporting Handshake to DOT
This commit transforms the ExportDOT pass into a tool to be more
consistent with the general behavior of existing passes (which
transform the IR) and tools (which may produce non-MLIR output).

The bulk of the DOT printer is moved to a support file that is
included from the straightforward export-dot tool, which is now built
by the project. The commit also brings some minor modifications to the
DOT exporter/printer to make its API simpler, more maintainable, and
more pleasing to clang-tidy.