[ExportDOT] Pretty-printing improvements
This commit slightly improves the node labels printed in the DOT
produced by the ExportDOT pass, making the visualization easier to look
at when not running the tool in debug mode. It also adds an mlir_op
attribute to each node to allow external tools to parse the produced
DOTs and match nodes to dataflow components more easily.