[ArithReduceStrength] Multiplier simplification
This commit introduces a new rewrite pattern to the ArithReduceStrength
pass that attempts to simplify multiplications with constants by
replacing them with a tree of adders/shifts. This can lead to better
performance and/or area.

To achieve this, we add support for modeling a semi-generic operation
tree in C++, which can then be "built" into a sequence of MLIR
operations. This will be useful in the future for similar
strength reduction-related optimizations.

At the moment, a pass parameter indicates the maximum depth of the adder
tree under which it makes sense to perform the replacememt. Eventually,
we could use the delay models from the buffer placement pass to drive
this parameter, as well as a general optimization target like area or
performance.