[ConcretizeIndexType] Verification functions
This commit adds a couple verification functions to the
HandshakeConcretizeIndexType pass that verify whether the operation they
are passed as argument (or any of its nested operations) has an
IndexType in its operands or results. After the pass is ran, all these
functions should succeed on any operation in the IR.

These functions are used as preconditions in a number of passes that do
not know how to handle IndexType's properly. As such, it is similar in
purpose to the verifyAllValuesHasOneUse function from CIRCT.