[CDGAnalysis] Support functions and test pass (#6)
This commit introduces support for control-data-graph (CDG) analysis,
which is a pre-requisite to the upcoming "FPL22 pass" (elastic pass with
fast-token delivery). The analysis is at this point experimental and
made up of a few support functions and data-structures used to derive
and represent the CDG of a standard-level function. 

The commit also includes the infrastructure to run tests on experimental
code as well as a test pass to check the behavior of the analysis using
FileCheck (by annotating some IR operations with the analysis results).

Co-authored-by: Lucas Ramirez <lucas.rami@proton.me>