[handshake-sim] JSON config for exec models (#11)
This commits adds support for providing a configuration file as well as
overrides to choose the set of operation execution models used during an
execution of the Handshake simulator. The JSON simply specifies a
mapping between operation name and name of the execution model to use
for it, while the --change-model command-line flag allows one to enforce
the use of specific execution models regardless of the configuration
present in the JSON.

At this point this is just preparatory work, as the configuration is
not used during simulation.