[HandshakeToNetlist] Improve module names (#10)
This improves the name given to external hardware modules in the
HandshakeToNetlist conversion pass. The derivation of the name of each
type of operation now depends on its specific discriminating parameters
(in the context of VHDL export). The goal of this preparatory work is to
make it easier to concretize VHDL modules in the upcoming export-vhdl
tool by encoding the discriminating parameters of each operation in a
regular way that is amenable for automated parsing.