[OptimizeBitwidths] New bitwidth optimization pass
This commit adds a new bitwidth optimization pass to Dynamatic++ which
is meant to be more complete and easily extensible than the one already
present. It relies solely on greedy rewriting, folding, and rewrite
patterns to implement the classical forward and backward passes required
to reduce the bitwidths of channels to a minimum. This new pass also
implements custom logic to optimize the bitwidths of all eligible
Handshake operations in addition to the traditional optimizations
applied on arithmetic operations. The pass and every one of its rewrite
pattern are heavily tested to ensure that all operations are handled
correctly.

Nevertheless, the pass currently has some limitations.
1. It is unable to optimize the bitwidth of float type values because we
   currently have no model of what floating widths we support during
   VHDL export.
2. Sometimes, inserted extension operations will be of the wrong type
   (arithmetic instead of logical, or the other way around), which can
   change the circuit semantics.
Future changes will address these limitations.