[OptimizeBitwidths] Correct ext. type placement
This makes the HandshakeOptimizeBitwidths pass place extension
operations of the correct type (arithmetic or logical) on operands and
results of operations being optimized, and replaces the incorrect
extension type determination strategy used before.

We achieve this by keeping track of the type(s) of extension operations
that are potentially being encountered whenever backtracking a value to
its minimal version. This tells us how a value should be interepreted
(signed or unsigned) and, consequently, how it should be treated when
the same semantic value is extended somewhere else in the IR.

Additionally, the commit fixes the rewrite pattern on return operations
to match on DynamaticReturnOp instead of on ReturnOp, which was the
intended behavior in the first place. This particular rewrite pattern is
also disabled because the optimization it performs conflicts with the
operation's verification function. As it may become useful in the
future, the pattern is kept in the code.