[OptimizeBitwidths] Bound optimization
Adds a rewrite pattern to the HandshakeBitwidthOptimization pass that
matches a stucture in the IR commonly used to make a control flow
decision (e.g., a loop condition) based on the comparison between a
number and a constant. Due to the complexity of the pattern, a new file
of unit tests is added to verify the pattern's behavior on simple and
complicated (nested loops) cases.