[OptimizeBitwidths] Forward cycle optimization (#20)
This introduces a new rewrite pattern that operates on merge-like
operations in the bitwidth optimization pass. The pattern optimizes
channels that are in "forwarding cycles" (i.e., sequences of Handshake
operations that forward one of their inputs to one of their outputs
without modification) and reduces their bitwidth based on the bitwidths
of all other values coming into the cycle through operands of merge-like
operations present in the cycle. On real benchmarks, this optimization
is very often necessary to reduce the bitwidth of channels that are part
of nested loops.

Fixes #14.