[Buffers] FPGA20 smart buffer placement (#8)
This commit adds a smart buffer placement pass to Dynamatic++, enabling
users to create high-performance dataflow circuits. The pass operates on
Handshake-level IR; it uses latency models of dataflow components as
well as user-specifiable per-channel buffering properties to build an
MILP which, when optimized, gives away a buffer placement strategy for
the circuit. The theory behind this buffer placement strategy is
described in the following paper published at FPGA'20:
https://www.epfl.ch/labs/lap/wp-content/uploads/2020/03/JosipovicFeb20_BufferPlacementAndSizingForHigh-PerformanceDataflowCircuits_FPGA20.pdf    

The latency models of all components are defined in JSON format in
data/components.json, and parsed using LLVM's JSON parsing library. We
use Gurobi's C++ API to describe and solve the MILP (which require users
to install Gurobi at the moment). In the future, we could open the
system to more solvers by having the pass produce a standard-formatted
text file containing the mathematical system instead of relying on a
proprietary API.