[ArithReduceStrength] Code organization
Light clean up and code moving in the arithmetic strength reduction
pass. Some static functions have become class methods when they are only
used as part of a single rewrite pattern and the implementation of
`OpTree` was moved at the top of the file to be more cleanly separated
from the rest.