[RotateForLoops] Support for negative bounds
This commit improves the loop rotation pass by letting it rewrite for
loops with negative bounds and which provably execute at least once. As
an implication of this change, the predicate of the comparison operation
inserted in the created do-while loop is now determined by the method
which assesses the legality of the rotation: it is signed if the loop's
lower bound + the step cannot be guaranteed to be non-negative, and
unsigned otherwise.

This commit also makes some cosmetic changes to the pass, generally
making its implementation cleaner.