[ArithReduceStrength] Fix incorrect cmp promotion
This changes effectively disables the signed integer comparison rewrite
pattern in the strength reduction pass. After examining our benchmarks,
it appears that MLIR/Polygeist allows negative values to be represented
with an IndexType, which violates the assumption that our promotion
pattern was making. As such, we cannot rely on the operand type to
perform comparison promotion. If we want to reinstate this optimizatiom,
we will have to define a more complicated heuristic to determine whether
the operands of a comparison are positive.