[IfToSelect] Convert simple if's to select's
Introduces a new pass (`ScfSimpleIfToSelect`, --scf-simple-if-to-select)
that converts `scf::IfOp` operations with a specific structure into
a combination of an `arith::SelectOp` and another arithmetic operation
that depends on the content of the if's branches. This has the nice
effect of removing a point of diverging control flow in the IR, hence
reducing the number of basic blocks at the std-level and below. For
example, it reduces by 2 the number of basic blocks in our
"binary_search" benchmark. Unit tests are also introduces for the pass.