[OptimizeBitwidths] Generalized bound optimization
This change improves significantly the bound optimization pattern of the
bitwidth optimization pass. It makes it able to reason about the
branch's condition operand when it is expressed as a conjunction of
integer comparison results. This makes the pattern strictly better and
more general. It benefits, for example, our "binary_search" benchmark
which features such constructs.

Another rewrite pattern is also introduced to simplify sequences of
extension operations followed by a truncation. This kind of IR structure
can arise as the result of the application of our other patterns and is
not handled automatically (at least not in all cases) by MLIR's existing
canonicalization/folding infrastructure.

Finally, an outdatded comment in the file header is removed.