[OptimizeBitwidths] More robust backtracking
This makes the backtracking and forwarding cycle detection logic in the
bitwidth optimization pass more robust and able to handle more
operations. It is now possible to backtrack through
`handshake::ConditionalBranchOp` and to detect forwarding cycle that
include `arith::SelectOp` operations, which behave live muxes, as well
as extension operations. It also fixes a typo in the forwarding cycle
detection logic that would make the function return false (i.e, no
cycle) when there was in fact a cycle. The latter fix opens more
optimization opportunities for the associated rewrite pattern.