[Buffers] Move CSV freq. parsing code to profiler
This change moves the code responsible for parsing the CSV file
produced by the experimental frequency profiler from the CFDFC
extraction code to the profiler itself, where it more naturally belongs.
The backing data structure (ArchBB) that holds the parsed information is
also moved for consistency. The buffer placement pass is also modified
as a consequence of this change and already sees some clean up related
to the tracking of selected archs and BBs during each CFDFC extraction
round.