[Buffers] Refactored buffer placement MILP
This commit continues the refactoring effort around the buffer placement
pass and focuses on defining a cleaner API for setting up and eventually
optimizing the large MILP that determines where to place buffers in our
dataflow circuits. Many parts of the code are moved around, remamed,
made more robust, etc.

Notably, we define a BufferPlacementMILP class that handles the entire
process, from setting up the MILP to extracting buffer placement
information from its solution after optimization. The goal of this class
is to allow future users and developers to quickly get an overview of
the entire process and potentially personnalize it for their needs by
subclassing the class and overriding/adding/removing some of the steps
involved in determining a buffer placement.

Documentation is currently sporadic in the newly defined entities but
will be added shortly in a future commit.