[Buffers] Heavy MILP refactoring and documentation
This continues the refactoring effort around the buffer placement pass.
In particular, this commit focuses on defining an even more flexible and
future-proof API for the buffer placement MILP and adds siginificantly
more documentation to the infrastructure, which is fairly complex and
stateful. When something is unclear or seems to diverge from the legacy
implementation, a /// NOTE is explicitly left in the code fir future
reference.

Further work includes better handling of channel-specific buffering
properties (which is extremely hacky right now), formalization of
placement result data types and logic, large refactoring of component
information parsing code, etc.