[Buffers] Better buffering properties API
This commit extends and improves the API to facilitate the manipulation
of channel-specific buffering properties. In particular, it is now
possible to attach properties to function arguments in the same way as
other channels.

Bumps CIRCT at the same time to get our latest local changes on the
ChanelBufProps struct.