[SetBufferingProperties] New buffer-related pass
This commit adds a pass that is meant to be an IR pre-processing step
before buffer placement. The pass iterates over all channels in a
Handshake function and sets channel-specific buffering properties
wherever necessary forllowing a predefined (and at this point, unique)
policy controlled by a pass parameter. This will be immediately useful
to set the buffering constraints in the IR for the first version of the
buffer placement pass (FPGA'20).