[TimingModels] Refactor of timing modeling infrastructure
This commit is a large refactoring of the infrastructure surrounding
timing models, which right now are only used by the buffer placement
pass. The main changes are:
1. Better and safer JSON parsing, which now follows the ADL-friendly API
   recommended in the LLVM documentation.
2. Cleaner data types to deserialize the JSON data to, for easier
   management, encapsulation, and access to component timing models.
3. Better documentation throughout the code.
4. TimingCharacterization[.h/.cpp] are renamed to TimingModels[.h/.cpp]
   to please the lazy-typers.

The parts of the buffer placement implementation that interact with
timing models are also refactored to make use of the new API. At the
same time, some improvements to error reporting are made to increase
reliability.

This commit largely concludes the effort to refactor and clean up the
FPGA'20 buffer placement pass (#16).