[InferBasicBlocks] Inference heuristic improvements
This commit significantly improves the inference heuristic at the core
of the HandshakeInferBasicBlocks pass to allow it to find the logical
basic block of operations in more complex situations. This is achieved
by having operations look at their predecessors (as before) AND
successors to find the block they logically belong to, and by having the
pass run the inference logic over all eligible operations inside
Handshake functions potentially multiple times (i.e., as long as there
is progress being made on inference).

More unit tests are added for the pass to evaluate the inference logic
in new situations where it did not apply before.