[HandshakeCanonicalize] Handshake canonicalization pass
This commit introduces a simple Handshake canonicalization pass that as
of now contains some simple rewrite patterns that aim to erase or
downgrade some Handshake operations in the IR. Even in this simple
form, the pass has the nice side-effect of making the resulting IR
compatible with legacy Dynamatic by (1) deleting unconditional branches
(which do not exist in legacy Dynamatic) and (2) erasing merge-like
operations with a single data operand (which are not always supported by
legacy Dynamatic). Unit tests are added to check the behavior of each
individual rewrite pattern.

In the future, more complex rewrite patterns will be implemented to
optimize our dataflow circuits even more aggressively. The pass's
existence rationale is explained in
lib/Transforms/HandshakeCanonicalize.cpp's comment header.