[OptimizeBitwidths] Do not reduce indices to i0
This commit adds two rewrite patterns to the bitwidth optimization pass
which ensure that the latter doesn't create i0 values when reducing the
index/select width of control merges and muxes, respectively (which was
previously possible when they had a single data operand). A unit test is
added for those new conversion patters.

This also adds support for displaying constants with 0 width in the DOT
printer (before this commit, it would assert when such a constant was
present). Prior to this change the bitwidth optimization pass could
indirectly lead to the creation such constants. Even if it no longer
happens, it costs us nothing to be able to display them.