[ReduceStrength] Reintroduce `arith::CmpIOp` promotion (#37)
This small change re-enables the comparison promotion rewrite pattern in
the ArithReduceStrength pass, which transforms signed integer
comparisons into corresponding unsigned ones whenever both of the
comparison's operands can be proven to be positive. Now that we have a
numeric analysis framework in place, we can use it to make such
promotion whenever possible. This also means that future improvements to
the analysis will directly benefit the strength reduction pass.

Fixes #26.