[handshake-sim] Support for Dynamatic operations (#32)
This commit is the first version of the handshake-simulator tool now
supporting Dynamatic operations.

- Added new Dynamatic operations : Load, Store, Return, End, Memory
Controller
- Added support for internal states. That way, operations can store data
that can be read by everyone at any cycle in a clear way
- Fixed major memory bugs from the default code
- Added a simple data structure for models arguments, allowing global
modifications very easily and making the code more readable
- Removed obsolete features, mostly because we aren't using func dialact
functions

The simulator is thus fully functional for translated handshake code for
Dynamatic, however cycle accuracy is not guaranteed by this PR. It
should be pretty cycle accurate at this point, but this is very unclear
and fully untested for now. Also, it lacks of many unit tests, which
have to be done right after this PR, meaning there has to be many bugs
for particular cases.