[export-dot] Improve printing mode handling
This commit improves the way that the export-dot tool handles various
printing modes, which each impact visual and/or structural aspects of
the generated DOT files. A new "mode" command line argument replaces the
now deleted "legacy" and "dot-debug" parameters, which didn't allow to
express the combinations of visual/structural features that make sense
with regard to the downstream uses we have for DOT files. Additionally,
and "edge-style" command line argument is added to allow control over
the way graph edges are styled when visualized in GraphViz.