[ScfToCf] Signedness-aware `scf::ForOp` lowering  (#38)
This introduces a new conversion pass to Dynamatic++ (--lower-scf-to-cf)
that is almost identical to the upstream `--convert-scf-to-cf` pass,
which, as its name indicates, converts `scf` operations to unstructured
control flow (`cf`). The only difference between the two is that the
`scf::ForOp` rewrite pattern used in our pass (which overwrites the
default one) inserts an _unsigned_ comparison operation for checking
whether the loop iterator is within the for loop's bounds instead of a
_signed_ one if it can prove that both the iterator and upper bound are
guaranteed to be positive. The estimation relies on the recently
introduced numerical analysis. This has a performance impact
down-the-line during bitwidth analysis.

Fixes #28.