[CFDFC] Backedges only come from branches
Slightly modifies the logic to determine which channels of a CFDFC are
backedges. In addition to the regular characteristics of a backedge, the
CFDFC only marks channels which originate from a branch-like operation
as backedges. This has an impact on the buffer placement MILP, which
uses the set of backedges identified by each CFDFC do decide where to
numerically insert "tokens" in the circuit.