[Buffers] Generalized channel executions counting
This changes (actually, reverts) the way we count the total number of
channel executions for buffer placement so that it is a more accurate
estimation of the "real" value. The previous logic also had limitations
with respect to the structure of the circuits that it could support.

Because of this change, this number is now marginally different from
what the legacy buffer pass obtains for the same computation, which can
(mostly) cosmetically impact buffer placement since the number is used
in the MILP objective.