[Test] Add first 9 integration tests (FPL22)
This commit adds an initial set of 9 integration tests (C kernels) to
Dynamatic which have been cleaned up and formalized from legacy
Dynamatic. Every integration test is made up of a header containing the
kernel's declaration, a source file containing the kernel's
implementation, and of a set of test inputs for the kernel which are
used for profiling before buffer placement. Formal documentation on how
the conventions that these files must obey will come in a future commit
when some progress has been made on the Dynamatic frontend.

These particular kernels were chosen because they are the ones which
were which were used in the evaluation of the fast-token paper (FPL'22).